refactor(ui): extract chrome widgets to ui/widgets/
Moves the four chrome renderers — tabs bar, filter input, status
bar, and the loading screen — into ui/widgets/, with one file
each. Marks panel_block as pub(crate) so child modules can reach
it. Total ~190 lines move out of mod.rs; call sites in draw() and
the snapshot tests stay unchanged via a use-statement.
