+++ src/ui/widgets/mod.rs
@@ -1,0 +1,8 @@
+//! Reusable chrome widgets that wrap the main content area:
+//! tabs bar at the top, filter input + status bar at the bottom,
+//! and the loading screen shown during startup.
+
+pub(super) mod filter_input;
+pub(super) mod loading;
+pub(super) mod status_bar;
+pub(super) mod tabs_bar;
